+++ app/dashboard/admin/equipment/[labId]/page.tsx
@@ -4,9 +4,9 @@ import { redirect } from 'next/navigation';
 import Link from 'next/link';
 import { PlusIcon, PencilIcon, TrashIcon } from '@heroicons/react/24/outline';
 
-export default async function LabEquipmentPage({ params }: { params: { labId: string } }) {
+export default async function LabEquipmentPage({ params }: { params: Promise<{ labId: string }> }) {
   const session = await auth();
-
+  
   // Check if user is authenticated and has admin role
   if (!session?.user || session.user.role !== 'ADMIN') {
     redirect('/dashboard');
