+++ app/dashboard/admin/labs/[labId]/delete/page.tsx
@@ -1,0 +1,8 @@
+import LabDeletePage from "@/app/_components/lab-delete-page";
+
+export default async function Page({params}: {params: Promise<{labId: string}>}) {
+  const {labId} = await params;
+  return (
+    <LabDeletePage labId={labId}/>
+  );
+}
