+++ app/dashboard/admin/labs/[labId]/page.tsx
@@ -1,0 +1,8 @@
+export default async function Page({params}: {params: Promise<{ labId: string }>}) {
+    const { labId } = await params;
+    return (
+        <>
+            <h1>Hello {labId}</h1>
+        </>
+    )
+}
