+++ prisma/schema.prisma
@@ -83,8 +83,8 @@ model Lab {
 model UserLab {
   userId       String
   labId        String
-  user         User        @relation(fields: [userId], references: [id])
-  lab          Lab         @relation(fields: [labId], references: [id])
+  user         User        @relation(fields: [userId], references: [id], onDelete: Cascade)
+  lab          Lab         @relation(fields: [labId], references: [id], onDelete: Cascade)
   @@id([userId, labId])
 }
 
@@ -95,5 +95,5 @@ model Equipment {
   description  String?
   labId        String
   quantity     Int
-  lab          Lab        @relation(fields: [labId], references: [id])
+  lab          Lab        @relation(fields: [labId], references: [id], onDelete: Cascade)
 }
